Move PornDetectorController to Rover
