+++ kifi-backend/modules/scraper/conf/scraperService.routes
@@ -5,9 +5,6 @@
 GET     /internal/scraper/status                        @com.keepit.scraper.ScraperController.status
 POST    /internal/scraper/getBasicArticle               @com.keepit.scraper.ScraperController.getBasicArticle
 POST    /internal/scraper/getSignature                  @com.keepit.scraper.ScraperController.getSignature
-GET     /internal/scraper/getPornDetectorModel          @com.keepit.controllers.scraper.PornDetectorController.getModel
-POST    /internal/scraper/pornDetector/detectPorn       @com.keepit.controllers.scraper.PornDetectorController.detect
-POST    /internal/scraper/pornDetector/whitelist        @com.keepit.controllers.scraper.PornDetectorController.whitelist
 
 ->  / commonService.Routes
